video_service/requirements.txt
Django==5.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
requests==2.31.0
kafka-python==2.0.2
python-decouple==3.8
Pillow==10.0.0
drf-yasg==1.21.7
